refactor event_binner into its own files
* event_binner definition removed from cell_group.hpp header
* event_binner is not templated, so implementation placed in
  cpp file
